// src/screens/user/types.tsx
export interface Geo {
  lat: string;
  lng: string;
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}

export interface ApiResponse {
  users: User[];
  totalPages: number;
  currentPage: number;
  hasMore: boolean;
}

export type RootStackParamList = {
  UserList: undefined;
  UserDetail: { user: User };
};